pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "python-dotenv>=1.2.2",
    "livekit-agents[silero,turn-detector]~=1.4",
    "livekit-plugins-noise-cancellation~=0.2",
    "pymongo>=4.16.0",
]
